moving main to root
